+++ app/src/main/java/me/saket/dank/urlparser/UrlParser.java
@@ -12,6 +12,7 @@ import net.dean.jraw.models.Submission;
 import java.util.Objects;
 import java.util.concurrent.ExecutionException;
 import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import javax.inject.Inject;
 import javax.inject.Named;
@@ -23,6 +24,7 @@ import me.saket.dank.utils.JrawUtils2;
 import me.saket.dank.utils.Optional;
 import me.saket.dank.utils.Urls;
 import okhttp3.HttpUrl;
+import timber.log.Timber;
 
 /**
  * Parses URLs found in the wilderness of Reddit and categorizes them into {@link Link} subclasses.
@@ -38,11 +40,13 @@ public class UrlParser {
 
   private final Cache<String, Link> cache;
   private final UrlParserConfig config;
+  private final Pattern imgurPreviewPat;
 
   @Inject
   public UrlParser(@Named("url_parser") Cache<String, Link> cache, UrlParserConfig config) {
     this.cache = cache;
     this.config = config;
+    this.imgurPreviewPat = Pattern.compile("_d(\\.\\w+)$");
   }
 
   /**
@@ -261,13 +265,23 @@ public class UrlParser {
         .scheme("https")
         .build();
 
-    // Attempt to get direct links to images from Imgur submissions.
-    // For example, convert 'http://imgur.com/djP1IZC' to 'http://i.imgur.com/djP1IZC.jpg'.
-    if (!isImageOrGifUrlPath(url) && !url.endsWith("mp4")) {
-      // If this happened to be a GIF submission, the user sadly will be forced to see it instead of its GIFV.
-      imageUrl = imageUrl.newBuilder(imageUrl.encodedPath() + ".jpg")
-          .host("i.imgur.com")
-          .build();
+    String imageUrlPath = imageUrl.encodedPath();
+
+    if (!url.endsWith("mp4")) {
+      if (isImageOrGifUrlPath(imageUrlPath)) {
+        // Strip any preview-related suffixes and queries
+        imageUrl = imageUrl.newBuilder()
+            .encodedQuery(null)
+            .encodedPath(imgurPreviewPat.matcher(imageUrlPath).replaceFirst("$1"))
+            .build();
+      } else {
+        // Attempt to get direct links to images from Imgur submissions.
+        // For example, convert 'http://imgur.com/djP1IZC' to 'http://i.imgur.com/djP1IZC.jpg'.
+        // If this happened to be a GIF submission, the user sadly will be forced to see it instead of its GIFV.
+        imageUrl = imageUrl.newBuilder(imageUrlPath + ".jpg")
+            .host("i.imgur.com")
+            .build();
+      }
     }
 
     //noinspection ConstantConditions
@@ -360,7 +374,8 @@ public class UrlParser {
   }
 
   public static boolean isImagePath(String urlPath) {
-    return urlPath.endsWith(".png") || urlPath.endsWith(".jpg") || urlPath.endsWith(".jpeg");
+    return urlPath.endsWith(".png") || urlPath.endsWith(".jpg") ||
+        urlPath.endsWith(".jpeg") || urlPath.endsWith(".webp");
   }
 
   private static boolean isImageOrGifUrlPath(String urlPath) {