+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - Fix playing of reddit hosted videos ([#210](https://github.com/Tunous/Dawn/pull/210))
 - Imgur images (up)loading ([#202](https://github.com/Tunous/Dawn/pull/202))
+- Always show fullsize imgur images ([#209](https://github.com/Tunous/Dawn/pull/209))
 
 ## [0.9.1] - 2020-05-05
 ### Fixed